Move PlatformPaths into Core (#839)
